+++ docs/helpers/index.js
@@ -1,7 +1,6 @@
 export { default as Paginator } from './Paginator';
 export { default as PerPage } from './PerPage';
 export { default as PrimaryControls } from './PrimaryControls';
-export { default as Search } from './Search';
 export { default as generateRows } from './generate_rows';
 export { default as paginate } from './paginate';
 export { default as VisibilityToggles } from './VisibilityToggles';