+++ packages/reactabular-search-field/tests/search_test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import TestUtils from 'react-addons-test-utils';
 import { expect } from 'chai';
-import { Search } from '../../docs/helpers/';
+import Search from '../src';
 
 describe('Search', function () {
   it('does not render options without rows', function () {