+++ packages/reactabular/CHANGELOG.md
@@ -1,3 +1,8 @@
+1.2.4 / 2016-08-08
+==================
+
+  * Improvement - search-columns - Rewrite as a stateless function. You should pass `query` as a prop now.
+
 1.2.3 / 2016-08-08
 ==================